MIFOSAC-278 migrated model package to the room related package
