+++ core/network/src/main/java/com/mifos/core/network/services/ChargeService.kt
@@ -10,11 +10,11 @@
 package com.mifos.core.network.services
 
 import com.mifos.core.entity.client.Charges
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.objects.clients.ChargeCreationResponse
 import com.mifos.core.objects.clients.Page
 import com.mifos.core.objects.template.client.ChargeTemplate
 import com.mifos.core.payloads.ChargesPayload
+import com.mifos.room.basemodel.APIEndPoint
 import okhttp3.ResponseBody
 import retrofit2.http.Body
 import retrofit2.http.GET
