+++ core/network/src/main/java/com/mifos/core/network/services/CheckerInboxService.kt
@@ -9,11 +9,11 @@
  */
 package com.mifos.core.network.services
 
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.network.GenericResponse
 import com.mifos.core.objects.checkerinboxtask.CheckerInboxSearchTemplate
 import com.mifos.core.objects.checkerinboxtask.CheckerTask
 import com.mifos.core.objects.checkerinboxtask.RescheduleLoansTask
+import com.mifos.room.basemodel.APIEndPoint
 import retrofit2.http.DELETE
 import retrofit2.http.GET
 import retrofit2.http.POST
