+++ core/network/src/main/java/com/mifos/core/network/services/ClientAccountsService.kt
@@ -9,7 +9,7 @@
  */
 package com.mifos.core.network.services
 
-import com.mifos.core.model.APIEndPoint
+import com.mifos.room.basemodel.APIEndPoint
 import com.mifos.room.entities.accounts.ClientAccounts
 import retrofit2.http.GET
 import retrofit2.http.Path
