+++ core/network/src/main/java/com/mifos/core/network/services/ClientService.kt
@@ -12,7 +12,6 @@ package com.mifos.core.network.services
 import com.mifos.core.entity.client.Client
 import com.mifos.core.entity.client.ClientPayload
 import com.mifos.core.entity.templates.clients.ClientsTemplate
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.network.GenericResponse
 import com.mifos.core.objects.clients.ActivatePayload
 import com.mifos.core.objects.clients.ClientAddressRequest
@@ -22,6 +21,7 @@ import com.mifos.core.objects.noncoreobjects.Identifier
 import com.mifos.core.objects.noncoreobjects.IdentifierCreationResponse
 import com.mifos.core.objects.noncoreobjects.IdentifierPayload
 import com.mifos.core.objects.noncoreobjects.IdentifierTemplate
+import com.mifos.room.basemodel.APIEndPoint
 import com.mifos.room.entities.accounts.ClientAccounts
 import okhttp3.MultipartBody
 import okhttp3.ResponseBody
