+++ core/network/src/main/java/com/mifos/core/network/services/CollectionSheetService.kt
@@ -9,11 +9,11 @@
  */
 package com.mifos.core.network.services
 
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.network.GenericResponse
 import com.mifos.core.network.model.IndividualCollectionSheetPayload
 import com.mifos.core.network.model.RequestCollectionSheetPayload
 import com.mifos.core.objects.collectionsheets.CollectionSheetRequestPayload
+import com.mifos.room.basemodel.APIEndPoint
 import com.mifos.room.entities.collectionsheet.CenterDetail
 import com.mifos.room.entities.collectionsheet.CollectionSheetPayload
 import com.mifos.room.entities.collectionsheet.CollectionSheetResponse
