+++ core/network/src/main/java/com/mifos/core/network/services/DocumentService.kt
@@ -9,9 +9,9 @@
  */
 package com.mifos.core.network.services
 
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.network.GenericResponse
 import com.mifos.core.objects.noncoreobjects.Document
+import com.mifos.room.basemodel.APIEndPoint
 import okhttp3.MultipartBody
 import okhttp3.ResponseBody
 import retrofit2.http.DELETE
