+++ core/network/src/main/java/com/mifos/core/network/services/GroupService.kt
@@ -11,11 +11,11 @@ package com.mifos.core.network.services
 
 import com.mifos.core.entity.group.Group
 import com.mifos.core.entity.group.GroupPayload
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.network.GenericResponse
 import com.mifos.core.objects.clients.ActivatePayload
 import com.mifos.core.objects.clients.Page
 import com.mifos.core.objects.responses.SaveResponse
+import com.mifos.room.basemodel.APIEndPoint
 import com.mifos.room.entities.accounts.GroupAccounts
 import com.mifos.room.entities.group.GroupWithAssociations
 import retrofit2.http.Body
