+++ core/network/src/main/java/com/mifos/core/network/services/LoanService.kt
@@ -14,7 +14,6 @@ import com.mifos.core.entity.accounts.loan.LoanWithAssociations
 import com.mifos.core.entity.accounts.loan.Loans
 import com.mifos.core.entity.client.Charges
 import com.mifos.core.entity.templates.loans.LoanRepaymentTemplate
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.network.GenericResponse
 import com.mifos.core.network.model.LoansPayload
 import com.mifos.core.objects.account.loan.LoanApproval
@@ -24,6 +23,7 @@ import com.mifos.core.objects.clients.Page
 import com.mifos.core.objects.organisations.LoanProducts
 import com.mifos.core.objects.template.loan.GroupLoanTemplate
 import com.mifos.core.payloads.GroupLoanPayload
+import com.mifos.room.basemodel.APIEndPoint
 import com.mifos.room.entities.templates.loans.LoanTemplate
 import com.mifos.room.entities.templates.loans.LoanTransactionTemplate
 import retrofit2.http.Body
