+++ core/network/src/main/java/com/mifos/core/network/services/SearchService.kt
@@ -9,8 +9,8 @@
  */
 package com.mifos.core.network.services
 
-import com.mifos.core.model.APIEndPoint
 import com.mifos.core.objects.SearchedEntity
+import com.mifos.room.basemodel.APIEndPoint
 import retrofit2.http.GET
 import retrofit2.http.Query
